requirements.txt
fastapi
uvicorn
sentence-transformers
faiss-cpu
pandas
numpy
together
better_profanity
